+++ src/main/java/org/torproject/metrics/stats/clients/Database.java
@@ -14,7 +14,6 @@ import java.time.Instant;
 import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.List;
-import java.util.Locale;
 import java.util.TimeZone;
 
 /** Database wrapper to connect to the database, insert data, run the stored
@@ -95,8 +94,7 @@ class Database implements AutoCloseable {
     String columns = "date, node, country, transport, version, frac, users";
     statistics.add(columns.split(", "));
     Statement st = this.connection.createStatement();
-    Calendar calendar = Calendar.getInstance(TimeZone.getTimeZone("UTC"),
-        Locale.US);
+    Calendar calendar = Calendar.getInstance(TimeZone.getTimeZone("UTC"));
     String queryString = "SELECT " + columns + " FROM estimated";
     try (ResultSet rs = st.executeQuery(queryString)) {
       while (rs.next()) {
@@ -120,8 +118,7 @@ class Database implements AutoCloseable {
     String columns = "date, node, country, transport, version, frac, low, high";
     statistics.add(columns.split(", "));
     Statement st = this.connection.createStatement();
-    Calendar calendar = Calendar.getInstance(TimeZone.getTimeZone("UTC"),
-        Locale.US);
+    Calendar calendar = Calendar.getInstance(TimeZone.getTimeZone("UTC"));
     String queryString = "SELECT " + columns + " FROM combined";
     try (ResultSet rs = st.executeQuery(queryString)) {
       while (rs.next()) {