+++ src/main/java/org/torproject/metrics/web/DirectoryListing.java
@@ -18,7 +18,6 @@ import java.net.URLConnection;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
-import java.util.Locale;
 import java.util.Map;
 import java.util.SortedSet;
 import java.util.TreeSet;
@@ -201,7 +200,7 @@ public class DirectoryListing extends HashMap<String, List<String[]>>
     int exp = (int) (Math.log(bytes) / Math.log(1024));
     char pre = "KMGTPE".charAt(exp - 1);
     return String
-        .format(Locale.US, "%.1f %siB", bytes / Math.pow(1024, exp), pre);
+        .format("%.1f %siB", bytes / Math.pow(1024, exp), pre);
   }
 }