+++ src/main/java/org/torproject/metrics/web/UpdateNews.java
@@ -14,6 +14,7 @@ import java.io.InputStreamReader;
 import java.net.URL;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Locale;
 
 public class UpdateNews {
 
@@ -21,6 +22,7 @@ public class UpdateNews {
 
   /** Update news. */
   public static void main(String[] args) throws Exception {
+    Locale.setDefault(Locale.US);
     URL textFile = new URL(
         "https://trac.torproject.org/projects/tor/wiki/doc/"
         + "MetricsTimeline?format=txt");