+++ CHANGELOG.md
@@ -9,6 +9,7 @@
    - Make Jetty host configurable.
    - Configure a base URL in order to turn ExoneraTor's permanent
      links into https:// links.
+   - Set default locale `US` at the beginning of the execution.
 
 
 # Changes in version 1.3.0 - 2019-11-09